requirements.txt
torch
flask
transformers
datasets
umap-learn